requirements.txt
pytest
pytest-json-report
pyyaml
streamlit
pandas